Removed nested authorization.
DETAILS:
With the change of authorization mechanism in Tastypie 0.9.12 that
allows finer grain authorization control, the current implementation of
checking authorization over nested resources becomes out-of-sync with
the new design.

Ultimately I think that it should be removed completely - as I don't
believe that the nested resource should have any difference in handling
authorization behaviour from the original resource.